+++ packages/virtual/react/src/components/Subscription/index.tsx
@@ -7,13 +7,10 @@ import type { SubscriptionProps } from "types";
  * React component.
  * Rerenders only on certain {@link @af-utils/virtual-core#(VirtualScrollerEvent:variable)}.
  * Allows to optimize performance.
- *
- * @privateRemarks
- * TODO: convert to arrow function when https://github.com/microsoft/rushstack/issues/1629 gets solved
  */
-function Subscription(props: SubscriptionProps): ReactNode {
+const Subscription = (props: SubscriptionProps): ReactNode => {
     useComponentSubscription(props.model, props.events);
     return props.children();
-}
+};
 
 export default Subscription;