+++ packages/virtual/react/src/hooks/useSubscription/index.ts
@@ -12,21 +12,16 @@ import { VirtualScroller, VirtualScrollerEvent } from "@af-utils/virtual-core";
  * @param model - {@link @af-utils/virtual-core#VirtualScroller} instance
  * @param events - array of {@link @af-utils/virtual-core#(VirtualScrollerEvent:variable)} to subscribe
  * @param callBack - callback to be called
- *
- * @privateRemarks
- * TODO: convert to arrow function when https://github.com/microsoft/rushstack/issues/1629 gets solved
  */
-function useSubscription(
+const useSubscription = (
     model: VirtualScroller,
     events: readonly VirtualScrollerEvent[] | VirtualScrollerEvent[],
     callBack: () => void
-) {
-    useEffect(() => {
+) => useEffect(() => {
         if (callBack) {
             callBack();
             return model.on(callBack, events);
         }
     }, [model, callBack, events]);
-}
 
 export default useSubscription;