+++ packages/virtual/react/src/hooks/useVirtual/index.ts
@@ -21,16 +21,13 @@ import type {
  *     overscanCount: 1
  * });
  * ```
- *
- * @privateRemarks
- * TODO: convert to arrow function when https://github.com/microsoft/rushstack/issues/1629 gets solved
  */
-function useVirtual(params: VirtualScrollerInitialParams): VirtualScroller {
+const useVirtual = (params: VirtualScrollerInitialParams): VirtualScroller => {
     const model = useVirtualModel(params);
 
     useIsomorphicLayoutEffect(() => model.set(params));
 
     return model;
-}
+};
 
 export default useVirtual;